+++ activerecord/test/cases/migration_test.rb
@@ -1700,7 +1700,7 @@ if ActiveRecord::Base.connection.supports_migrations?
       @migrations_path = MIGRATIONS_ROOT + "/valid"
       @existing_migrations = Dir[@migrations_path + "/*.rb"]
 
-      sources = ActiveSupport::OrderedHash.new
+      sources = {}
       sources[:bukkits] = MIGRATIONS_ROOT + "/to_copy"
       sources[:omg] = MIGRATIONS_ROOT + "/to_copy2"
       ActiveRecord::Migration.copy(@migrations_path, sources)
@@ -1741,7 +1741,7 @@ if ActiveRecord::Base.connection.supports_migrations?
       @migrations_path = MIGRATIONS_ROOT + "/valid_with_timestamps"
       @existing_migrations = Dir[@migrations_path + "/*.rb"]
 
-      sources = ActiveSupport::OrderedHash.new
+      sources = {}
       sources[:bukkits] = MIGRATIONS_ROOT + "/to_copy_with_timestamps"
       sources[:omg]     = MIGRATIONS_ROOT + "/to_copy_with_timestamps2"
 
@@ -1783,7 +1783,7 @@ if ActiveRecord::Base.connection.supports_migrations?
       @migrations_path = MIGRATIONS_ROOT + "/valid_with_timestamps"
       @existing_migrations = Dir[@migrations_path + "/*.rb"]
 
-      sources = ActiveSupport::OrderedHash.new
+      sources = {}
       sources[:bukkits] = MIGRATIONS_ROOT + "/to_copy_with_timestamps"
       sources[:omg]     = MIGRATIONS_ROOT + "/to_copy_with_name_collision"
 
@@ -1802,7 +1802,7 @@ if ActiveRecord::Base.connection.supports_migrations?
       @migrations_path = MIGRATIONS_ROOT + "/valid_with_timestamps"
       @existing_migrations = Dir[@migrations_path + "/*.rb"]
 
-      sources = ActiveSupport::OrderedHash.new
+      sources = {}
       sources[:bukkits] = MIGRATIONS_ROOT + "/to_copy_with_timestamps"
 
       skipped = []